enable pickling of agents and handlers
