+++ server/tasks/process_file.py
@@ -4,22 +4,24 @@ import pickle
 import os
 import logging
 
+from adala.agents import Agent
+
 from aiokafka import AIOKafkaConsumer
 from celery import Celery, states
 from celery.exceptions import Ignore
-from server.utils import get_input_topic, get_output_topic, ResultHandler, Settings
+from server.utils import get_input_topic, get_output_topic, Settings
+from server.handlers.result_handlers import ResultHandler
 
 
 logger = logging.getLogger(__name__)
 
 REDIS_URL = os.getenv("REDIS_URL", "redis://localhost:6379/0")
-app = Celery("worker", broker=REDIS_URL, backend=REDIS_URL)
+app = Celery("worker", broker=REDIS_URL, backend=REDIS_URL, accept_content=["json", "pickle"])
 
 
-@app.task(name="process_file", track_started=True)
-def process_file(serialized_agent: bytes):
-    # Load the agent
-    agent = pickle.loads(serialized_agent)
+@app.task(name="process_file", track_started=True, serializer='pickle')
+def process_file(agent: Agent):
+
     # # Read data from a file and send it to the Kafka input topic
     asyncio.run(agent.environment.initialize())
 
@@ -30,10 +32,8 @@ def process_file(serialized_agent: bytes):
     asyncio.run(agent.environment.finalize())
 
 
-@app.task(name="process_file_streaming", track_started=True, bind=True)
-def process_file_streaming(self, serialized_agent: bytes):
-    # Load the agent
-    agent = pickle.loads(serialized_agent)
+@app.task(name="process_file_streaming", track_started=True, bind=True, serializer='pickle')
+def process_file_streaming(self, agent: Agent):
 
     # Get own job ID to set Consumer topic accordingly
     job_id = self.request.id
@@ -45,22 +45,22 @@ def process_file_streaming(self, serialized_agent: bytes):
 
 
 async def async_process_streaming_output(
-    input_job_id: str, serialized_result_handler: bytes, batch_size: int
+    input_job_id: str, result_handler: ResultHandler, batch_size: int
 ):
     logger.info(f"Polling for results {input_job_id=}")
 
 
-    try:
-        import server.utils as utils
-        # result_handler = ResultHandler.__dict__[result_handler]
-        result_handler = pickle.loads(serialized_result_handler)
-        # assert isinstance(result_handler, ResultHandler)
-    except Exception as e:
-        from celery.contrib import rdb
-        rdb.set_trace()
-        # logger.error(f"{result_handler} is not a valid ResultHandler")
-        logger.error(f"not a valid ResultHandler")
-        raise e
+    # try:
+        # import server.utils as utils
+        # # result_handler = ResultHandler.__dict__[result_handler]
+        # result_handler = pickle.loads(serialized_result_handler)
+        # # assert isinstance(result_handler, ResultHandler)
+    # except Exception as e:
+    # from celery.contrib import rdb
+    # rdb.set_trace()
+        # # logger.error(f"{result_handler} is not a valid ResultHandler")
+        # logger.error(f"not a valid ResultHandler")
+        # raise e
 
     topic = get_output_topic(input_job_id)
     settings = Settings()
@@ -82,8 +82,11 @@ async def async_process_streaming_output(
             for tp, messages in data.items():
                 if messages:
                     result_handler(messages)
+                    logger.info(f"Handled {len(messages)} messages in topic {tp.topic}")
                 else:
                     logger.info(f"No messages in topic {tp.topic}")
+            if not data:
+                logger.info(f"No messages in any topic")
         finally:
             job = process_file_streaming.AsyncResult(input_job_id)
             if job.status in ["SUCCESS", "FAILURE", "REVOKED"]:
@@ -95,15 +98,17 @@ async def async_process_streaming_output(
     await consumer.stop()
 
 
-@app.task(name="process_streaming_output", track_started=True, bind=True)
+@app.task(name="process_streaming_output", track_started=True, bind=True, serializer='pickle')
 def process_streaming_output(
-    self, job_id: str, serialized_result_handler: bytes, batch_size: int = 2
+    self, job_id: str, result_handler: ResultHandler, batch_size: int = 2
 ):
     try:
-        asyncio.run(async_process_streaming_output(job_id, serialized_result_handler, batch_size))
+        asyncio.run(async_process_streaming_output(job_id, result_handler, batch_size))
     except Exception as e:
         # Set own status to failure
         self.update_state(state=states.FAILURE)
+
+        logger.log(level=logging.ERROR, msg=e)
 
         # Ignore the task so no other state is recorded
         # TODO check if this updates state to Ignored, or keeps Failed
